requirements.txt
Flask
flask_cors
Pillow
ultralytics
opencv-python
numpy
gunicorn